Clean AnalysisModal
