+++ src/components/data-management/AnalysisModal.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import {
   Modal,
@@ -11,22 +11,31 @@ import {
   List,
 } from 'antd';
 import { ClipLoader } from 'react-spinners';
+import moment from 'moment';
+import { useRouter } from 'next/router';
 import EditableField from '../EditableField';
 import { updateExperiment, saveExperiment } from '../../redux/actions/experiments';
-import validateInputs, { rules } from '../../utils/validateInputs';
+import {
+  updateProject,
+} from '../../redux/actions/projects';
+
+import { runGem2s } from '../../redux/actions/pipeline';
 
 const { Title } = Typography;
 
-const NewExperimentModal = (props) => {
+const AnalysisModal = (props) => {
   const {
     visible,
     onLaunch,
     onCancel,
-    activeProject,
-    experiments,
   } = props;
 
   const dispatch = useDispatch();
+  const router = useRouter();
+
+  const experiments = useSelector((state) => state.experiments);
+  const { activeProjectUuid } = useSelector((state) => state.projects.meta) || false;
+  const activeProject = useSelector((state) => state.projects[activeProjectUuid]);
 
   const [experimentsList, setExperimentsList] = useState([]);
   const [numFieldsEditing, setNumFieldsEditing] = useState(0);
@@ -45,10 +54,19 @@ const NewExperimentModal = (props) => {
     setIsWorking(!visible);
   }, [visible]);
 
-  const validationChecks = [
-    rules.MIN_1_CHAR,
-    rules.ALPHANUM_DASH_SPACE,
-  ];
+  const onLaunchAnalysis = (experimentId) => {
+    console.log('Launching this analysis');
+    setIsWorking(true);
+    onLaunch(experimentId);
+
+    const analysisPath = '/experiments/[experimentId]/data-processing';
+    const lastViewed = moment().toISOString();
+    dispatch(updateExperiment(experimentId, { lastViewed }));
+    dispatch(updateProject(activeProjectUuid, { lastAnalyzed: lastViewed }));
+
+    dispatch(runGem2s(experimentId));
+    router.push(analysisPath.replace('[experimentId]', experimentId));
+  };
 
   const renderAnalysisList = () => {
     if (!experimentsList?.length > 0) {
@@ -74,16 +92,9 @@ const NewExperimentModal = (props) => {
                   <Button
                     type='primary'
                     onClick={() => {
-                      setIsWorking(true);
-                      onLaunch(experiment.id);
+                      onLaunchAnalysis(experiment.id);
                     }}
                     disabled={numFieldsEditing > 0 || isWorking}
-                    onKeyDown={(e) => {
-                      if (e.key === 'Enter') {
-                        setIsWorking(true);
-                        onLaunch(experiment.id);
-                      }
-                    }}
                   >
                     Launch
                   </Button>
@@ -98,7 +109,6 @@ const NewExperimentModal = (props) => {
                     dispatch(updateExperiment(experiment.id, { name: name.trim() }));
                   }}
                   value={experiment.name}
-                  validationFunc={(name) => validateInputs(name, validationChecks).isValid}
                   deleteEnabled={false}
                   onEditing={(editing) => {
                     setNumFieldsEditing(Math.max(0, numFieldsEditing + (editing || -1)));
@@ -148,18 +158,16 @@ const NewExperimentModal = (props) => {
   );
 };
 
-NewExperimentModal.propTypes = {
+AnalysisModal.propTypes = {
   visible: PropTypes.bool,
   onCancel: PropTypes.func,
   onLaunch: PropTypes.func,
-  activeProject: PropTypes.object.isRequired,
-  experiments: PropTypes.object.isRequired,
 };
 
-NewExperimentModal.defaultProps = {
+AnalysisModal.defaultProps = {
   visible: true,
   onCancel: null,
   onLaunch: null,
 };
 
-export default NewExperimentModal;
+export default AnalysisModal;
